refactor: move software details to shared components
