+++ packages/shared-components/src/api/HttpRequest.ts
@@ -26,12 +26,19 @@ export class HttpRequest {
   public constructor(config: AxiosRequestConfig) {
     this.service = axios.create(config);
     this.service.interceptors.request.use(
-      (internalConfig: InternalAxiosRequestConfig) => internalConfig,
-      (error: AxiosError) => Promise.reject(error),
+      (config: InternalAxiosRequestConfig) => {
+        return config;
+      },
+      (error: AxiosError) => {
+        return Promise.reject(error);
+      },
     );
     this.service.interceptors.response.use(
       (response: AxiosResponse) => {
         const { data } = response;
+        if (data instanceof Blob) {
+          return data;
+        }
         if (data.code === Code.HANDLE_SEPARATELY) {
           return Promise.reject(data);
         }
@@ -56,8 +63,11 @@ export class HttpRequest {
   get<T>(url: string, params?: object): Promise<ResultData<T>> {
     return this.service.get(url, params);
   }
-
-  post<T>(url: string, params?: object, config?: AxiosRequestConfig): Promise<ResultData<T>> {
+  post<T>(
+    url: string,
+    params?: object,
+    config?: AxiosRequestConfig,
+  ): Promise<T extends Blob ? Blob : ResultData<T>> {
     return this.service.post(url, params, config);
   }
 }
