+++ packages/shared-components/src/components/index.ts
@@ -5,3 +5,4 @@ export { default as SearchSoftware } from './SearchSoftware.vue';
 export { default as TechLandscape } from './TechLandscape.vue';
 export { default as ProjectsCompare } from './ProjectsCompare.vue';
 export { default as CompareFavorites } from './CompareFavorites.vue';
+export { default as SoftwareDetails } from './SoftwareDetails.vue';
