+++ packages/website/src/router/index.ts
@@ -1,6 +1,8 @@
+import type { RouteRecordRaw } from 'vue-router';
 import { createRouter, createWebHashHistory } from 'vue-router';
+import { ElMessage } from 'element-plus';
 
-const routes = [
+const routes: RouteRecordRaw[] = [
   {
     path: '/',
     component: () => import('@views/Home.vue'),
@@ -8,6 +10,12 @@ const routes = [
   {
     path: '/software-details',
     component: () => import('@views/SoftwareDetails.vue'),
+    beforeEnter: to => {
+      if (!to.query.repoName) {
+        ElMessage.error('缺少 URL 参数');
+        return false;
+      }
+    },
   },
   {
     path: '/tech-landscape',
